100 - Move otel_data_model.py
