+++ tel2puml/otel_to_pv/data_holders/base.py
@@ -6,7 +6,7 @@ from typing import Self, Optional, Any, Generator
 import logging
 
 
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import OTelEvent
+from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import OTelEvent
 
 
 LOGGER = logging.getLogger(__name__)
