+++ tel2puml/otel_to_pv/data_sources/base.py
@@ -1,7 +1,7 @@
 """Module containing base classes to provide interfaces for data sources."""
 from abc import ABC, abstractmethod
 from typing import Self
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
+from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
     OTelEvent,
 )
