+++ tel2puml/otel_to_pv/data_sources/json_data_source/json_datasource.py
@@ -4,7 +4,7 @@ from typing import Any, Iterator
 
 import ijson
 
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
+from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
     OTelEvent,
     JSONDataSourceConfig,
 )
