+++ tests/tel2puml/otel_to_pv/data_sources/json_data_source/conftest.py
@@ -1,7 +1,7 @@
 """Configuration for tests in json data source."""
 import pytest
 
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
+from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
     FieldSpec,
 )
 from tel2puml.otel_to_pv.data_sources.json_data_source.json_jq_converter \
