+++ tests/tel2puml/otel_to_pv/otel_ingestion/test_otel_data_model.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 import pytest
 from pydantic import ValidationError
 
-from tel2puml.otel_to_pv.otel_ingestion.otel_data_model import (
+from tel2puml.otel_to_pv.data_holders.sql_data_holder.data_model import (
     load_config_from_dict
 )
